pyproject.toml
[build-system]
requires = [
    "poetry==1.1.13",
]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "quickstart_simulation"
version = "0.1.0"
description = "Federated Learning Simulation Quickstart with Flower"
authors = ["The Flower Authors <enquiries@flower.dev>"]

[tool.poetry.dependencies]
python = "^3.6.2"
flwr = { extras = ["simulation"], version = "^0.17.0" }
# flwr = { extras = ["simulation"], path = "../../", develop = true }  # Development
tensorflow-cpu = "2.6.2"
tensorflow-model-optimization = "0.7.2"
pandas = "1.0.2"